Added section about the SQLContainer in the API reference
Updated the docstrings of the sql_container module to support reStructuredText
and follow the numpy docstring style guide.

Added code examples to the docstrings of the SQLContainer.replace_placeholders method and
the Placeholder class.

Removed inheritance from abc.ABC for the SQLContainer class. It does not implement any abstract methods.

Did not include the method SQLContainer.validate_class_variables in the documentation yet since it is only
useful when more than one database dialect is implemented.