Renamed SQLite test database
Renamed the SQLite test database to match the name used in the documentation (Runescape.db).
When clicking the download link of the test database in the User guide of the documentation the actual filename
will be "Runescape.db" and not "RSGeneralStore.db". Runescape.db is a better name.

Updated conftest.py to use the new database names in the fixtures "sqlite_db_file" and "sqlite_db_file_empty".

Rename: RSGeneralStore.db -> Runescape.db and RSGeneralStore_empty.db -> Runescape_empty.db