Renamed "an SQL" to "a SQL"
Updated all occurrences of sentences with "an SQL" to "a SQL" with regex.
Also capitalized "SQL" everywhere.

The convention seems to be to say "a SeeQuElle" and not "an EssQuuElle" database.
This is also how pandas writes about SQL in their documentation.

Regex   : ([Aa])[Nn]([ \n\r])?[Ss][Qq][Ll]
Replace : $1$2SQL