Updated section about Exceptions in the API reference
Changed generating the Exceptions from the .. automodule:: directive to an .. autoexception:: directive
for each Exception. This makes the module name "exceptions" not appear in the Exception name in the generated output.
E.g. ExecuteStatementError should be displayed as pandemy.ExecuteStatementError not pandemy.exceptions.ExecuteStatementError
since all pandemy features are exposed at the package level.