Updated __str__ and __repr__ methods
Updated __str__ method of SQLiteDb to display the file attribute
enclosed with '', which looks better.

Updated  __repr__ of DatabaseManager to use the repr format string convention flag !r
on each attribute set on the DatabaseManager. This creates a better debugging
representation of the object.

Updated the unit tests and the documentation to reflect these changes.