Updated cross references of the documentation.
Added and updated cross references in the documentation.

Objects not from pandemy are now consistently prefixed with the foreign library
e.g. pandas|sqlalchemy.some_method_or_function().

Methods and functions now display a () in the reference to indicate that they are callable.
E.g. DatabaseManager.save_df().

Pandemy objects are not prefixed with pandemy to reduce verbosity. However methods are now prefixed with
the class they belong to for increased clarity.

Pandemy Exceptions are still prefixed with pandemy to distinguish them better from built-in
Exceptions.