Updated docstrings for DatabaseManager and SQLiteDb
The numpy style guide says:

"An Attributes section, located below the Parameters section, may be used to describe non-method attributes of the class"

Removed `container` and `engine_config` from the Attributes section since they are set from the arguments to the
initializer.

Only `conn_str` and `engine` are now listed as attributes since they are created during the initialization.

Removed the Attributes section from SQLiteDb to avoid repetition since it is listed in its superclass DatabaseManager.