Updated setup.cfg
- Updated the metadata.
- Added install_requires for pandas and SQLAlchemy.
- Added section options.extras_require for development dependencies of pytest and sphinx.

Updated pandemy.__version__ to be a simple version string and not be generated from pandemy.__versiontuple__
to avoid errors when building the package. pandemy.__version__ also cannot be typed with "str" as
this also causes errors during the build process.

The package is now installable (with both required and optional dependencies) with pip and soon ready for release:
$ pip install -f .\dist\ Pandemy
or
$ pip install -f .\dist\ "Pandemy[tests, docs]"