Added function df_to_parameters_in_chunks
This helper function of module _dataframe converts a DataFrame into an
iterator yielding a list of dicts, where each dict is a row of the DataFrame.
The function should be used for adding the chunksize parameter to the methods
upsert_table and merge_df.

Added unit tests for df_to_parameters_in_chunks.