Updated the save_df if_exists option
Fixed bug for if_exists='replace'. It now correctly deletes the data of the existing table
before writing the DataFrame instead of dropping the table, recreating it, and finally writing the DataFrame.

Added option 'drop-replace' to the if_exists parameter. This option does what the bug with 'replace' did before,
which is also the way if_exists='replace' is implemented in the pandas.DataFrame.to_sql method.

Added validation of the chunksize parameter.

Added and updated unit tests to cover the changes.