Updated DatabaseManager
Refactored DatabaseManager from an ABC into a proper class.
Implemented the __init__ method. The DatabaseManager can now be initialized
from a SQLAlchemy URL or Engine. This is useful if you want to use Pandemy
but there is no subclass of DatabaseManager implemented for the desired SQL
dialect. Added unit tests for the __init__, __str__ and __repr__ methods.