Reworked the internals of SQLiteDb
Simplified the __init__ method and removed the helper methods
_set_attributes, _build_conn_str and _create_engine as they are no longer needed.
Their work are replaced by the call to super().__init__.

Added the new parameters to the SQLiteDb.__init__: driver, url, connect_args and engine.

The driver parameter decides what database driver to use. The default is
the built-in module sqlite3.

Removed unnecessary validation of the parameters container and engine_config to promote
duck typing.

Updated the unit tests of the __init__ method.