Reworked OracleDb.__init__
Simplified the __init__ method and added support for providing a url or an engine
to initialize OracleDb. This removes the need for the alternative constructors from_url and
from_engine and aligns with how __init__ works on DatabaseManager and SQLiteDb.
Added call to super().__init__.

Added the driver parameter to decide what database driver to use. The default is cx_oracle.

Updated and added unit tests of the __init__ method.