Updated function pandemy._datetime.localize_and_convert_timezone
Changed the behaviour to raise pandemy.InvalidInputError if supplying
a naive datetime column and a value for `target_tz` instead of silently
ignoring the timezone conversion. It is not possible to convert the timezone
of a naive datetime column and it is better to notify the user of this
through an exception instead of skipping the conversion.