Removed validation of the `df` and `conn` parameters of DatabaseManager.save_df
Validation that the `df` and `conn` parameters are an instance of pandas.DataFrame and
sqlalchemy.engine.Connection respectively makes it hard to use subclasses that behave like a
DataFrame or a Connection, e.g. mocking the connection in a unit test.
Removing the validation is more in line with duck typing. Removed unit tests that send invalid
input to the parameters.