Updated sections in the API Reference
Removed sub-sections for the DatabaseManager and SQLContainer in the API Reference chapter.
Sphinx 5.2.0 added domain objects to the table of contents, which makes it possible to see the
methods of e.g. DatabaseManager in the table of contents. Removing the sub-sections makes the
table of contents look better and giving a maximum of 3 levels of indentation instead of 4.